+++ Nbm/src/test/resources/projForTest/src/completion/paramInfoArg2.kt
@@ -1,0 +1,7 @@
+package completion
+
+fun paramInfoTarget2(x: Int, y: String = "hi", vararg z: Boolean) {}
+
+fun testParamInfoArg2() {
+    paramInfoTarget2(42, <caret>"hi", true)
+}